+++ filament/backend/src/opengl/platforms/PlatformEGLHeadless.cpp
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-#include "PlatformEGLHeadless.h"
+#include <backend/platforms/PlatformEGLHeadless.h>
 
-#include "opengl/OpenGLDriver.h"
-#include "opengl/OpenGLContext.h"
-#include "opengl/OpenGLDriverFactory.h"
+#include "opengl/GLUtils.h"
 
 #include <EGL/egl.h>
 #include <EGL/eglext.h>
@@ -47,7 +45,8 @@ PlatformEGLHeadless::PlatformEGLHeadless() noexcept
         : PlatformEGL() {
 }
 
-backend::Driver* PlatformEGLHeadless::createDriver(void* sharedContext, const Platform::DriverConfig& driverConfig) noexcept {
+backend::Driver* PlatformEGLHeadless::createDriver(void* sharedContext,
+        const Platform::DriverConfig& driverConfig) noexcept {
     EGLBoolean bindAPI = eglBindAPI(EGL_OPENGL_API);
     if (UTILS_UNLIKELY(!bindAPI)) {
         slog.e << "eglBindAPI EGL_OPENGL_API failed" << io::endl;
@@ -199,7 +198,7 @@ backend::Driver* PlatformEGLHeadless::createDriver(void* sharedContext, const Pl
     clearGlError();
 
     // success!!
-    return OpenGLDriverFactory::create(this, sharedContext, driverConfig);
+    return OpenGLPlatform::createDefaultDriver(this, sharedContext, driverConfig);
 
 error:
     // if we're here, we've failed